[Spring Boot][David] test: add 404 not found for invald company id when getting specific company in intergration test

+++ src/main/java/com/thoughtworks/springbootemployee/Service/CompanyService.java
@@ -6,7 +6,9 @@ import com.thoughtworks.springbootemployee.Repository.CompanyRepository1;
 import com.thoughtworks.springbootemployee.Repository.EmployeeRepository;
 import com.thoughtworks.springbootemployee.Repository.EmployeeRepository1;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -31,11 +33,16 @@ public class CompanyService {
     }
 
     public Company getById(String id) {
-        return companyRepository.findById(id).orElse(null);
+        try{
+            return companyRepository.findById(id).orElse(null);
+        }
+        catch(Exception exception){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Company Not Found!");
+        }
     }
 
     public List<Employee> getEmployeesByCompanyId(String id) {
-        Company company =  companyRepository.findById(id).orElse(null);
+        Company company =  getById(id);
         if(company != null){
             List<String> employeeIds = company.getEmployeesId();
             Iterable<Employee> employees = employeeRepository.findAllById(employeeIds);
@@ -58,7 +65,7 @@ public class CompanyService {
     }
 
     public Company update(String id, Company companyUpdate) {
-        if (companyRepository.findById(id).orElse(null) != null) {
+        if (getById(id) != null) {
             companyUpdate.setId(id);
             return companyRepository.save(companyUpdate);
         }
@@ -66,6 +73,8 @@ public class CompanyService {
     }
 
     public void delete(String id) {
-        companyRepository.deleteById(id);
+        if (getById(id) != null) {
+            companyRepository.deleteById(id);
+        }
     }
 }
